Refactor: 폴더구조 변경 파일과 폴더명에 kebab-case
