+++ src/components/common/slider/slider.tsx
