+++ src/components/common/slider/style.ts
