+++ src/components/common/solve-status.tsx
