+++ src/components/common/user-info/user-solved-info.tsx
