+++ src/pages/member/[detail].tsx
@@ -1,15 +1,15 @@
 import styled from "styled-components";
-import LineChart from "@/components/chart/chart";
-import UserInfo from "@/components/userInfo/UserInfo";
-import UserSolvedInfo from "@/components/userInfo/UserSolvedInfo";
-import Tab from "@/components/tab/Tab";
-import SolvedRecord from "@/components/tab/SolvedRecord";
+import LineChart from "@/components/common/chart/chart";
+import UserInfo from "@/components/common/user-info/user-info";
+import UserSolvedInfo from "@/components/common/user-info/user-solved-info";
+import Tab from "@/components/common/tab/tab";
+import SolvedRecord from "@/components/common/tab/solved-record";
 import Board from "@/components/common/board/Board";
 import { useSelector } from "react-redux";
 import { studyApi } from "@/api/studyApi";
 import { useRouter } from "next/router";
 import { memberApi } from "@/api/memberApi";
-import Loading from "@/components/Loading/Loading";
+import Loading from "@/components/common/loading/Loading";
 import { GetServerSideProps } from "next";
 import { parseCookies } from "@/util/parseCookie";
