+++ src/pages/rank/study.tsx
@@ -1,7 +1,7 @@
 import styled from "styled-components";
 import Board from "@/components/common/board/Board";
 import { studyApi } from "@/api/studyApi";
-import Loading from "@/components/Loading/Loading";
+import Loading from "@/components/common/loading/Loading";
 
 const StudyRank = () => {
   const {data, isLoading} = studyApi.useGetAllStudyListQuery({})
