+++ src/pages/study/[studyId]/index.tsx
@@ -1,14 +1,14 @@
 import styled from "styled-components";
-import StudyInfo from "@/components/study/studyInfo";
-import Tab from "@/components/tab/Tab";
+import StudyInfo from "@/components/study/study-info";
+import Tab from "@/components/common/tab/tab";
 import Board from "@/components/common/board/Board";
-import SolvedRecord from "@/components/tab/SolvedRecord";
-import LineChart from "@/components/chart/chart";
+import SolvedRecord from "@/components/common/tab/solved-record";
+import LineChart from "@/components/common/chart/chart";
 import { useSelector } from "react-redux";
-import SolveStatus from "@/components/common/solveStatus";
+import SolveStatus from "@/components/common/solve-status";
 import { studyApi } from "@/api/studyApi";
 import { useRouter } from "next/router";
-import Loading from "@/components/Loading/Loading";
+import Loading from "@/components/common/loading/Loading";
 import { parseCookies } from "@/util/parseCookie";
 import { GetServerSideProps } from "next";
 import { useState, useEffect } from "react";
