+++ src/pages/study/manage.tsx
@@ -1,6 +1,6 @@
 import styled from "styled-components";
-import Input from "@/components/common/Input";
-import Slider from "@/components/slider/slider";
+import Input from "@/components/common/input";
+import Slider from "@/components/common/slider/slider";
 import useInput from "@/hooks/useInput";
 import { useEffect } from "react";
 import { useRouter } from "next/router";
